feat:update room after finish
