+++ src/dataBase/dataBaseGames.ts
@@ -1,3 +1,3 @@
 import { Ship } from "../utils/types";
 
-export const dataBaseGames = new Map<number, Map<number, { indexPlayer: string, ships: Ship[], killedShips: number }>>();
+export const dataBaseGames = new Map<number, Map<number, { indexPlayer: string, ships: Ship[], killedShips: number, moves: { x: number, y: number }[] }>>();
