+++ src/webSocket_server/game_process/attack_handler.ts
@@ -15,49 +15,56 @@ export const attackHandler = (req: Request) => {
 
   if (nextPlayer.index === indexPlayer) {
 
-    let status: string = 'miss';
-    const indexRival = indexPlayer === 0 ? 1 : 0;
-    let emptyPositions = [] as { x: number; y: number; }[];
+    //const move = { x: xPosition, y: yPosition };
+    const g = dataBaseGames.get(gameId)?.get(indexPlayer)?.moves.find((move) => move.x === xPosition && move.y === yPosition);
+    if (g === undefined) {
+      dataBaseGames.get(gameId)?.get(indexPlayer)?.moves.push({ x: xPosition, y: yPosition })
+      let status: string = 'miss';
+      const indexRival = indexPlayer === 0 ? 1 : 0;
+      let emptyPositions = [] as { x: number; y: number; }[];
 
-    for (const ship of dataBaseGames.get(gameId)!.get(indexRival)!.ships) {
+      for (const ship of dataBaseGames.get(gameId)!.get(indexRival)!.ships) {
 
-      for (const position of ship.positionEachDeck!) {
-        if (position[1].x === xPosition && position[1].y === yPosition) {
+        for (const position of ship.positionEachDeck!) {
+          if (position[1].x === xPosition && position[1].y === yPosition) {
 
-          attackedShip.push({ x: xPosition, y: yPosition, direction: ship.direction });
-          ship.positionEachDeck!.delete(position[0]);
+            attackedShip.push({ x: xPosition, y: yPosition, direction: ship.direction });
+            ship.positionEachDeck!.delete(position[0]);
 
-          if (ship.positionEachDeck!.size === 0) {
-            status = 'killed';
+            if (ship.positionEachDeck!.size === 0) {
+              status = 'killed';
 
-            const ammountKilledShips = dataBaseGames.get(gameId)!.get(indexRival)!.killedShips;
-            dataBaseGames.get(gameId)!.get(indexRival)!.killedShips = ammountKilledShips + 1;
+              const ammountKilledShips = dataBaseGames.get(gameId)!.get(indexRival)!.killedShips;
+              dataBaseGames.get(gameId)!.get(indexRival)!.killedShips = ammountKilledShips + 1;
 
-            emptyPositions = setEmptyPositionAround(attackedShip);
-            attackedShip.length = 0;
+              emptyPositions = setEmptyPositionAround(attackedShip);
+              attackedShip.length = 0;
 
+            } else {
+              status = 'shot';
+            }
+            nextPlayer.index = indexPlayer;
+            break;
           } else {
-            status = 'shot';
+            nextPlayer.index = indexPlayer === 0 ? 1 : 0;
           }
-          nextPlayer.index = indexPlayer;
-          break;
-        } else {
-          nextPlayer.index = indexPlayer === 0 ? 1 : 0;
         }
+        if (status !== 'miss') break;
       }
-      if (status !== 'miss') break;
-    }
-    sendResponseAfterAttack(status, xPosition, yPosition, indexPlayer, gameId, nextPlayer.index);
+      sendResponseAfterAttack(status, xPosition, yPosition, indexPlayer, gameId, nextPlayer.index);
 
-    if (status === 'killed') {
-      emptyPositions.forEach((position) => {
-        sendResponseAfterAttack("miss", position.x, position.y, indexPlayer, gameId, nextPlayer.index)
-      })
-    }
+      if (status === 'killed') {
+        emptyPositions.forEach((position) => {
+          sendResponseAfterAttack("miss", position.x, position.y, indexPlayer, gameId, nextPlayer.index)
+        })
+      }
 
-    if (dataBaseGames.get(gameId)!.get(indexRival)!.killedShips === 10) {
-      finishGame(gameId, indexPlayer);
+      if (dataBaseGames.get(gameId)!.get(indexRival)!.killedShips === 10) {
+        finishGame(gameId, indexPlayer);
 
+      }
+    } else {
+      console.log("There has already been such a move")
     }
   } else {
     console.log("Next player's turn")
