+++ src/webSocket_server/game_process/finish.ts
@@ -1,9 +1,11 @@
 import { wss } from "../../..";
 import { connections } from "../../dataBase/dataBaseConnections";
 import { dataBaseGames } from "../../dataBase/dataBaseGames";
+import { updateWinners } from "./update_winners";
+import { updateRoom } from "./update_room";
 
 export const finishGame = (gameId: number, indexPlayer: number) => {
-  console.log('heello')
+
   const resp = {
     type: "finish",
     data: JSON.stringify({
@@ -11,19 +13,7 @@ export const finishGame = (gameId: number, indexPlayer: number) => {
     }),
     id: 0
   }
-  const nameWinnerIndex = dataBaseGames.get(gameId)?.get(indexPlayer)?.indexPlayer;
-  const nameWinner = connections.get(nameWinnerIndex!)!.name;
-  const wins = connections.get(nameWinnerIndex!)!.wins;
-  connections.get(nameWinnerIndex!)!.wins = wins + 1;
-
-  const resp2 = {
-    type: "update_winners",
-    data: JSON.stringify([{
-      name: nameWinner,
-      wins: connections.get(nameWinnerIndex!)!.wins,
-    }],),
-    id: 0,
-  }
+  wss.clients.forEach((client) => client.send(JSON.stringify(resp)));
 
   const firstPlayer = dataBaseGames.get(gameId)!.get(0)!.indexPlayer;
   const secondPlayer = dataBaseGames.get(gameId)!.get(1)!.indexPlayer;
@@ -34,7 +24,9 @@ export const finishGame = (gameId: number, indexPlayer: number) => {
 
   if (connections.has(secondPlayer)) {
     connections.get(secondPlayer)?.ws.send(JSON.stringify(resp))
-  }
-  wss.clients.forEach((client) => client.send(JSON.stringify(resp2)));
+  };
+
+  updateWinners(gameId, indexPlayer);
+  updateRoom();
 
 }
